Install zsh-plugins via package manager

+++ zsh/completion.zsh
@@ -9,18 +9,20 @@
 # | General |
 # +---------+
 
-# source ./gambit.zsh
-
 # Remove the crappy zsh completion shipped by Homebrew git.
 # I prefer zsh's builtin git completion.
 # https://github.com/Homebrew/homebrew-core/issues/33275
-[[ -f  $HOMEBREW_PREFIX/share/zsh/site-functions/_git ]] && \
+if [[ -f  $HOMEBREW_PREFIX/share/zsh/site-functions/_git ]]; then
   rm -f $HOMEBREW_PREFIX/share/zsh/site-functions/_git
+fi
 
 # Load more completions
-fpath=($DOTFILES/zsh/plugins/zsh-completions/src $fpath)
+# fpath=($DOTFILES/zsh/plugins/zsh-completions/src $fpath)
 # Homebrew also brings some completions
-fpath=($HOMEBREW_PREFIX/share/zsh/site-functions $fpath)
+if [[ -d $HOMEBREW_PREFIX/share/zsh/site-functions ]]; then
+  fpath=($HOMEBREW_PREFIX/share/zsh/site-functions $fpath)
+fi
+
 # Load my custom completions
 fpath=($DOTFILES/zsh/completions $fpath)
